Moved battle system into a function. Started on implementing main manus
