+++ ArshLukeMarko_Group_13/ArshLukeMarko_Group_13.vcxproj
@@ -127,13 +127,15 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Display.c" />
     <ClCompile Include="GameFunction.c" />
     <ClCompile Include="Main.c" />
-    <ClCompile Include="Utility.c" />
+    <ClCompile Include="Interface.c" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Display.h" />
     <ClInclude Include="GameFunction.h" />
-    <ClInclude Include="Utility.h" />
+    <ClInclude Include="Interface.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
