+++ ArshLukeMarko_Group_13/GameFunction.c
@@ -1,83 +1,4 @@
 #include "GameFunction.h"
-#include "Utility.h"
-
-//Print Monsters
-//Players
-void printNibbles(bool status)
-{
-	printf("\n(\\___/)\n");
-	if (status == true)
-		printf("(='.'=)\n");
-	else
-		printf("(=x.x=)\n");
-	printf("(\")_(\")\n");
-}
-void printFelix(bool status)
-{
-	printf("\n/\\__/\\\n");
-	if (status == true)
-		printf("(=^.^=)\n");
-	else
-		printf("(=x.x=)\n");
-	printf("(\")_(\")_/\n");
-}
-void printPip(bool status)
-{
-	printf("\n  ___\n");
-	if (status == true)
-		printf(" (o>o)\n");
-	else
-		printf(" (x>x)\n");
-	printf("((___))\n");
-	printf("  ^ ^\n");
-}
-void printBaconBoi(bool status)
-{
-	printf("\n /\\_/\\\n");
-	if (status == true)
-		printf("( o.o )\n");
-	else
-		printf("( x.x )\n");
-	printf("( u u )\n");
-}
-//Opponents
-void printChomp(bool status)
-{
-	printf("\n		(\\___/)\n");
-	if (status == true)
-		printf("		( o_o )\n");
-	else
-		printf("		( x_x )\n");
-	printf("		(\")_(\")\n");
-}
-void printTy(bool status)
-{
-	printf("\n		/\\__/\\\n");
-	if (status == true)
-		printf("		(=0_0=)\n");
-	else
-		printf("		(=x_x=)\n");
-	printf("		(\")_(\")_/\n");
-}
-void printJared(bool status)
-{
-	printf("\n		 _____\n");
-	if (status == true)
-		printf("		( ovo )\n");
-	else
-		printf("		( xvx )\n");
-	printf("		((___))\n");
-	printf("		  ^ ^\n");
-}
-void printPorkchop(bool status)
-{
-	printf("\n		 /\\_/\\\n");
-	if (status == true)
-		printf("		( 'o' )\n");
-	else
-		printf("		( xox )\n");
-	printf("		( v v )\n");
-}
 
 double attackDamage(MONSTSTAT attacker, MONSTSTAT defender)
 {
@@ -97,56 +18,97 @@ void delay(int ms) //freezes the program for some number of milliseconds
 	while (clock() < start_time + ms);
 }
 
-void printBattleState(MONSTSTAT player, MONSTSTAT opponent, int p ,  int o) //displays the current state to the player
+void battleStart(MONSTSTAT player, MONSTSTAT opponent, int p, int o, double d)
 {
-	system("@cls||clear");
-	//print enemy
-	if (opponent.HP >= 0)
-		printf("\nOpponent's HP: %d / %d", (int)ceil(opponent.HP), (int)ceil(opponent.maxHP));
-	else
-		printf("\nOpponent's HP: 0 / %d", (int)ceil(opponent.maxHP));
-	bool monsterIsAlive;
-	if (opponent.HP > 0)
-		monsterIsAlive = true;
-	else
-		monsterIsAlive = false;
-	switch (o)
+	//Setup
+	opponent.attack = opponent.attack * d;
+	opponent.defence = opponent.defence * d; //Makes opponent's stats harder or easier depending on difficulty choice
+	//Start
+	int moveChoice;
+	double damage;
+	srand(time(NULL));
+	while (player.HP > 0 && opponent.HP > 0) //Battle starts
 	{
-	case 4:
-		printChomp(monsterIsAlive);
-		break;
-	case 5:
-		printTy(monsterIsAlive);
-		break;
-	case 6:
-		printJared(monsterIsAlive);
-		break;
-	case 7:
-		printPorkchop(monsterIsAlive);
-		break;
+		printBattleState(player, opponent, p, o); //print battle state
+		//Player's turn ---------------------------------------
+		printf("Select a move :\n1. Attack\n2. Defend\n3. Special\n4. Forfeit\n"); //print move options
+		if (scanf("%d", &moveChoice))
+		{
+			switch (moveChoice) //move choice logic
+			{
+			case 1: //Basic attack
+				damage = attackDamage(player, opponent);
+				opponent.HP = opponent.HP - damage;
+				printBattleState(player, opponent, p, o);
+				printf("You attacked your opponent for %.1lf damage.\n", damage);
+				break;
+			case 2: //Defence 
+				player.defence = player.defence * DEFENCE_MULTIPLIER;
+				printBattleState(player, opponent, p, o);
+				printf("Defence increased by 20%%\n");
+				break;
+			case 3: //Special move
+				damage = specialDamage(player, opponent);
+				if (damage > 0)
+					opponent.HP = opponent.HP - damage;
+				else
+					player.HP = player.HP + damage; //positive because damage is already negative in this case.
+				printBattleState(player, opponent, p, o);
+				if (damage > 0)
+					printf("Critical hit! You dealt %.1lf damage to your opponent!\n", damage);
+				else
+					printf("It backfired! You dealt %.1lf damage to yourself!\n", -damage);
+				break;
+			case 4: //Forfeit
+				player.HP = 0;
+				break;
+			}
+		}
+		//Enemy's turn ----------------------------------------
+		if (player.HP > 0 && opponent.HP > 0)
+		{
+			printf("\nOpponent's turn");
+			for (int i = 0; i < 5; i++)
+			{
+				delay(DELAY_TIME);
+				printf(".");
+			}
+			moveChoice = rand() % 3 + 1; //randomly select move
+			switch (moveChoice) //logic for player's move choice
+			{
+			case 1: //Basic attack
+				damage = attackDamage(opponent, player);
+				player.HP = player.HP - damage;
+				printBattleState(player, opponent, p, o);
+				printf("Your opponent attacked you for %.1lf damage.", damage);
+				break;
+			case 2: //Defence 
+				player.defence = opponent.defence * DEFENCE_MULTIPLIER;
+				printBattleState(player, opponent, p, o);
+				printf("Opponent's defence increased by 20%%");
+				break;
+			case 3: //Special move
+				damage = specialDamage(opponent, player);
+				if (damage > 0)
+					player.HP = player.HP - damage;
+				else
+					opponent.HP = opponent.HP + damage; //positive because damage is already negative in this case.
+				printBattleState(player, opponent, p, o);
+				if (damage > 0)
+					printf("Critical hit! Your opponent dealt %.1lf damage to you!", damage);
+				else
+					printf("Your opponents attack backfired! It dealt %.1lf damage to itself!", -damage);
+				break;
+			}
+			for (int i = 0; i < 5; i++)
+			{
+				printf(".");
+				delay(DELAY_TIME);
+			}
+		}
 	}
-	//print player
-	if (player.HP > 0)
-		monsterIsAlive = true;
+	if (player.HP <= 0) //end game logic, with trophies
+		printf("\nBattle over. You lost.\n");
 	else
-		monsterIsAlive = false;
-	switch (p)
-	{
-	case 0:
-		printNibbles(monsterIsAlive);
-		break;
-	case 1:
-		printFelix(monsterIsAlive);
-		break;
-	case 2:
-		printPip(monsterIsAlive);
-		break;
-	case 3:
-		printBaconBoi(monsterIsAlive);
-		break;
-	}	
-	if (player.HP >=0)
-		printf("Your HP: %d / %d\n\n", (int)ceil(player.HP), (int)ceil(player.maxHP));
-	else
-		printf("Your HP: 0 / %d\n\n", (int)ceil(player.maxHP));
+		printf("\nBattle over. You won!\n");
 }
