+++ ArshLukeMarko_Group_13/Interface.h
@@ -2,5 +2,5 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
-#include <string.h>
-#include <time.h>
+
+int mainMenu(void);
