--- ArshLukeMarko_Group_13/Utility.c
@@ -1,2 +1,0 @@
-#include "GameFunction.h"
-#include "Utility.h"
