+++ ArshLukeMarko_Group_13/stats
@@ -4,15 +4,15 @@
 0.75
 85
 85
-13
+15
 0.9
 60
 60
-25
+23
 1
 100
 100
-10
+13
 0.65
 70
 70
